feat(ui): refactor App around scene hosts (Refs #106)
